% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{MBformat}
\alias{MBformat}
\title{Format multi-species data for multi-site occupancy}
\usage{
MBformat(comm, meta, tax, id.vars, group.vars, zeroes = NULL, states = NULL)
}
\arguments{
\item{comm}{Matrix containing all samples from all states}

\item{meta}{Metadata table containing ID variables and covariates}

\item{tax}{Taxonomy table containing parsed ASV taxonomic information.}

\item{id.vars}{Character string of sample ID information. For multi-state analysis, there shold be one ID that represents a sample from each state.}

\item{group.vars}{Character vector of covariates in metadata to be regressed against.}

\item{zeroes}{Minimum number of samples from which a species can be absent. If NULL, then species may be absent from all samples.}

\item{states}{Character string of the metadata column holding the sample state information.}
}
\description{
A short description...
}
\details{
The community matrix should already be in binary format by the time it gets here.
}
\author{
Jose Miguel Ponciano Castellanos (\code{josemi@ufl.edu}), Kara J.M. Taylor (\code{k.taylor2@ufl.edu})
}
