% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_models.R
\name{MBOcc}
\alias{MBOcc}
\title{Apply occupancy model to formatted MB list}
\usage{
MBOcc(MBOcc.obj, formulae, assign.psi)
}
\arguments{
\item{MBOcc.obj}{Formatted object output by \code{format} function.}

\item{formulae}{List of formulae to run for each psi.
Should be a list of lists of length equal to \code{assign.psi}.}

\item{assign.psi}{List of numeric strings indicating how many psi to estimate.
List should be of length equal to \code{formulae}.}
}
\description{
A short description...
}
\author{
Jose Miguel Ponciano Castellanos (josemi@ufl.edu), Kara J.M. Taylor (\code{k.taylor2@ufl.edu})
}
