% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_predMBOcc.R
\name{MBpredplot}
\alias{MBpredplot}
\title{Plot predictions from predict.MBOcc.obj.est}
\usage{
MBpredplot(predMBOcc, covs, facets, return = T)
}
\arguments{
\item{predMBOcc}{Output list from \code{pred.MBOcc} containing newdata and predicted values}

\item{covs}{List of covariates in the output models}

\item{facets}{List of faceting variables in the output models}

\item{return}{Boolean; return all \code{ggplot} as list}
}
\description{
Visualize predictions
}
\author{
Kara J.M. Taylor (\code{k.taylor2@ufl.edu})
}
