% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timestamp.R
\name{MBruntime}
\alias{MBruntime}
\title{Estimate time to model completion}
\usage{
MBruntime(n.sp, n.form, n.psi)
}
\arguments{
\item{n.sp}{The number of species}

\item{n.form}{The number of formulae}

\item{n.psi}{The number of unique psi assignments}
}
\description{
Estimate time to model completion
}
