% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mbocc.R
\name{plot}
\alias{plot}
\title{Plot MBOcc object}
\usage{
plot(MBOcc, plot = c("MLE", "Site", "State"), return = F)
}
\arguments{
\item{MBOcc}{MBOcc object}

\item{plot}{Which parameter estimates to plot, one of \code{MLE}, \code{Site}, or \code{State}}
}
\description{
Plotting function for MBOcc object
}
\author{
Kara J.M. Taylor (\code{k.taylor2@ufl.edu})
}
