% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_MBOcc.R
\name{predict.MBOcc.obj.est}
\alias{predict.MBOcc.obj.est}
\title{Predict from MBOcc object}
\usage{
\method{predict}{MBOcc.obj.est}(MBOcc.obj.est, newdata, level = 0.95)
}
\arguments{
\item{MBOcc.obj.est}{Output list from \code{MBOcc} containing estimates and best-fit models}

\item{newdata}{Dataframe containing all covariates pertaining to MBOcc model}

\item{level}{Confidence interval 1-alpha}
}
\description{
Obtain 95\% confidence intervals around predicted values
}
\author{
Kara J.M. Taylor (\code{k.taylor2@ufl.edu})
}
