% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummarizeCommTax.R
\name{SummarizeCommTax}
\alias{SummarizeCommTax}
\title{Summarize count data by higher taxon}
\usage{
SummarizeCommTax(comm, tax, level = NULL)
}
\arguments{
\item{comm}{community matrix}

\item{tax}{full matrix containing 7 taxonomic levels (out from SILVA reference)}

\item{level}{column of traits matrix containing taxonomic name to summarize to}
}
\description{
Adds up counts of taxon columns by some higher taxon identification. The second data frame should provide a key for mapping the column IDs to higher taxon names.
}
