% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constrained_model.R
\name{const.model}
\alias{const.model}
\title{Occupancy estimation for pathogen in four tissue types}
\usage{
const.model(formula, data, assign.psi)
}
\arguments{
\item{formula}{of the form {~ Age * Species} to designate the covariate columns of interest}

\item{data}{the data, containing the covariates and the frequencies of state occupation}

\item{assign.psi}{a vector of the same length as the set of state positions (here CINT (4)) to indicate how the tissues should be grouped}
}
\description{
A short description...
}
\author{
Jose Miguel Ponciano Castellanos (josemi@ufl.edu)
}
