% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstorderchange.R
\name{f.o}
\alias{f.o}
\title{Recode timeseries community table to first order sequential change}
\usage{
f.o(comm, id, sitevar, timevar, sampleID)
}
\arguments{
\item{comm}{Community matrix}

\item{id}{ID variable indicating unit of repeated measurements}

\item{sitevar}{vector indicating which samples represent time-repeated measures}

\item{timevar}{vector indicating the time step at which the samples were taken}

\item{sampleID}{vector indicating the sample IDs}
}
\description{
These functions facilitate data carpentry for community table and metadata.
Based on the abundance at time \code{t}, recode the abundance at time \code{t+1}:
\verb{N_\{t+1\} > N_\{t\} = 1} or \verb{N_\{t+1\} < N_\{t\} = 0}
}
\examples{
comm <- apply(comm, 2, f.o, sitevar=meta$site, timevar=meta$time)
}
\keyword{first-order}
\keyword{sequence}
