% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{format}
\alias{format}
\title{Format multi-species data for multi-site occupancy}
\usage{
format(comm, meta, tax, zeroes, states = NULL)
}
\arguments{
\item{zeroes}{Minimum number of samples from which a species can be absent.}
}
\description{
A short description...
}
\author{
Jose Miguel Ponciano Castellanos (\code{josemi@ufl.edu}), Kara J.M. Taylor (\code{k.taylor2@ufl.edu})
}
