% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupColors.R
\name{group.colors}
\alias{group.colors}
\title{Group color palette by taxonomy}
\usage{
group.colors(taxa)
}
\arguments{
\item{taxa}{Character string of taxon names as given in MBOcc.obj}
}
\description{
Generate a rainbow color palette for taxa, grouped by genus or lower.
}
\author{
Kara J.M. Taylor (\code{k.taylor2@ufl.edu})
}
