% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mbocc.R
\name{MBplot}
\alias{MBplot}
\title{Plot MBOcc object}
\usage{
MBplot(
  MBOcc,
  plot = c("MLE", "Site", "State"),
  return = F,
  odds = ifelse(plot == "MLE", T, NULL)
)
}
\arguments{
\item{MBOcc}{MBOcc object}

\item{plot}{Which parameter estimates to plot, one of \code{MLE}, \code{Site}, or \code{State}}

\item{return}{logical; return to device (FALSE, the default)? Or return to object (TRUE)?}

\item{odds}{logical; return MLE plot as log-odds (TRUE, the default)? Or as probability (FALSE)?}
}
\description{
Plotting function for MBOcc object
}
\author{
Kara J.M. Taylor (\code{k.taylor2@ufl.edu})
}
