% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recombine.R
\name{recombine}
\alias{recombine}
\title{Combine lists with the same names to append into new sub-lists}
\usage{
recombine(list1, list2, FUN, recursive = F)
}
\arguments{
\item{list1}{List to match to}

\item{list2}{List to match}

\item{FUN}{Appending function, e.g. \code{append}, \code{c}, \code{rbind}, \code{cbind}}

\item{recursive}{Logical; Should list elements also be checked for concatenation?}
}
\description{
Match lists by name and append
}
\author{
Kara J.M. Taylor (\code{k.taylor2@ufl.edu})
}
