% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abvbel.R
\name{u.d}
\alias{u.d}
\title{Recode community table to above (1) or below (0) species mean}
\usage{
u.d(x)
}
\arguments{
\item{x}{To be used with \code{apply}. Single element of community table}
}
\description{
These functions facilitate data carpentry for community table and metadata.
}
\examples{
comm <- apply(comm, 2, u.d)
}
\keyword{above-or-below-mean}
