% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unroll_data.R
\name{unroll}
\alias{unroll}
\title{Unroll MBOcc estimates from model output}
\usage{
unroll(MBOcc.obj, MBOcc.est, group.vars, site.list)
}
\description{
Format MBOcc output into table format for plotting
}
\author{
Kara J.M. Taylor (\code{k.taylor2@ufl.edu})
}
